requirements.txt
halo
openai
pyyaml